backend/requirements.txt
fastapi[all]
sqlalchemy
psycopg2-binary
python-jose[cryptography]
pandas
scikit-learn
lightgbm
shap
passlib==1.7.4
bcrypt==3.2.0
Faker